added: script download from server, active script handling, change watcher; fixed some bugs
