+++ src/store/actions.js
@@ -34,7 +34,7 @@ import {fetchAll as fetchAllFolders, create as createFolder, markFolderRead} fro
 import {deleteMessage, fetchEnvelopes, fetchMessage, setEnvelopeFlag, syncEnvelopes} from '../service/MessageService'
 import {showNewMessagesNotification} from '../service/NotificationService'
 import {parseUid} from '../util/EnvelopeUidParser'
-import {getScripts} from '../service/FiltersService'
+import {getScripts, getScript} from '../service/FiltersService'
 
 export default {
 	savePreference({commit, getters}, {key, value}) {
@@ -487,25 +487,52 @@ export default {
 				throw err
 			})
 	},
-	fetchSieveScripts({getters, commit}, accountID){
-		//const scripts = getScripts(accountID)
-		const scripts = [{
-			"name": "Filter_1",
-			"raw":
-`require ["fileinto", "test"];
-# rule:[test]
-if header :contains "subject" "\"bla\""
-{
-        fileinto "Drafts";
-}`,
-		}]
-
-		for (const script of scripts) {
+	async fetchSieveScripts({getters, commit}, accountID){
+		commit("setSieveLoading", true)
+		commit("resetFilterAccount", {accountID})
+		let scripts = await getScripts(accountID)
+		scripts.entries = scripts.entries.filter(x => x !== "nextcloud_includes")
+		if (scripts.active === "nextcloud_includes") {
+			const includes_raw = await getScript(accountID, "nextcloud_includes")
+			const includeRegex = /include\s*:personal\s*"(\S*)";/gm
+			let match = includeRegex.exec(includes_raw)
+			while (match !== null) {
+				if (scripts.entries.indexOf(match[1]) !== -1) {
+					const raw = await getScript(accountID, match[1])
+					commit("newFilterSet", {
+						"accountID": accountID,
+						"name": match[1],
+						"original_name": match[1],
+						"raw": raw,
+						"active": true,
+						"changed": false,
+					})
+				}
+				scripts.entries = scripts.entries.filter(x => x !== match[1])
+				match = includeRegex.exec(includes_raw)
+			}
+		} else {
+			const raw = await getScript(accountID, scripts.active)
 			commit("newFilterSet", {
 				"accountID": accountID,
-				"name": script.name,
-				"raw": script.raw,
+				"name": scripts.active,
+				"original_name": scripts.active,
+				"raw": raw,
+				"active": true,
+				"changed": false,
+			})
+			scripts.entries = scripts.entries.filter(x => x !== scripts.active)
+		}
+		for (const script of scripts.entries) {
+			const raw = await getScript(accountID, script)
+			commit("newFilterSet", {
+				"accountID": accountID,
+				"name": script,
+				"original_name": script,
+				"raw": raw,
+				"changed": false,
 			})
 		}
+		commit("setSieveLoading", false)
 	}
 }
